pyproject.toml
[tool.poetry]
name = "gnews"
version = "0.1.0"
description = ""
authors = ["ReyterYT <reyteryt5@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = [
    {version = "0.14.0", extras = ["standard"], platform = "linux"},
    {version = "0.14.0", platform = "*"}
]
fastapi = "0.68.0"
"discord.py" = {git = "https://github.com/Rapptz/discord.py"}
humanize = "3.11.0"
python-dotenv = "0.19.0"
jishaku = "2.2.0"
psutil = "5.8.0"
replit = "3.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
